apps/workers/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
scipy==1.11.4

# Data processing
pandas==2.1.4
pillow==10.1.0

# Communication
nats-py==2.6.0
redis==5.0.1
asyncpg==0.29.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
